Training with MNIST Dataset

+++ src/Layer.h
@@ -6,14 +6,14 @@ using namespace std;
 class Layer
 {
     public:
-        void setInput(vector<float> Input){input = Input;}
-        void setOutput(vector<float> Output){output = Output;}
+        void setInput(vector<double> Input){input = Input;}
+        void setOutput(vector<double> Output){output = Output;}
         virtual void feedforward(Layer *){}
         virtual void calculateGradient(Layer *){}
-        virtual void calculateGradient(vector<float> &){}
+        virtual void calculateGradient(vector<double> &){}
         virtual void updateWeight(){}
-        vector<float> input;
-        vector<float> output;
-        vector<float> delta;
-        vector<vector<float>> weights;
+        vector<double> input;
+        vector<double> output;
+        vector<double> delta;
+        vector<vector<double>> weights;
 };
